% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{add_entry}
\alias{add_entry}
\title{Add entry}
\usage{
add_entry(text, ...)
}
\arguments{
\item{text}{Starting text string}

\item{...}{Contents of the new entry. They will be pasted together with a space. Any vectors will be separated with a semicolon. Any # will be replaced by a division line}
}
\value{
Updated text variable with the new entry
}
\description{
Adds an entry to a string of text to be witten to pdf
}
