% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{before_after}
\alias{before_after}
\title{Before after}
\usage{
before_after(
  a = 1,
  max_iter = 1000,
  initial_dims = 100,
  pca = TRUE,
  perplexity = 35,
  eta = 100,
  theta = 0.5,
  centers = 3,
  normalization = "None",
  only_pca = T
)
}
\value{
None
}
\description{
Generate before after pdf with plots
}
