% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{centroid2peakmatrix}
\alias{centroid2peakmatrix}
\title{Exponential Decay Similarity}
\usage{
centroid2peakmatrix(file_path)
}
\arguments{
\item{a}{Vector a}

\item{b}{Vector b}

\item{method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{normalize}{Boolean value indicating whether the a and b vectors should be normalixzed}
}
\value{
Similarity between a and b
}
\description{
Determines the degree of similarity between a vector a and a vector b as 1/exp(d) where d corresponds to the distance between the two normalized vectors calculated according to the specified method
}
