% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_module.R
\name{compute_s1}
\alias{compute_s1}
\title{Compute S1}
\usage{
compute_s1(calculated_magnitude, experimental_magnitude, similarity_method)
}
\arguments{
\item{calculated_magnitude}{Calculated magnitude vector}

\item{experimental_magnitude}{Experimental magnitude vector}

\item{similarity_method}{Similarity method to be used}
}
\value{
Spectral similarity score S1
}
\description{
Returns the spectral similarity score (S1)
}
