% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_module.R
\name{compute_s2}
\alias{compute_s2}
\title{Compute S2}
\usage{
compute_s2(calculated_magnitude, image_correl)
}
\arguments{
\item{calculated_magnitude}{Calculated magnitude vector}

\item{image_correl}{Image correlation matrix of the experimental magnitude vector}
}
\value{
Spectral similarity score S2
}
\description{
Returns the intra-cluster morphological similarity score (S2)
}
