% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_module.R
\name{compute_s3}
\alias{compute_s3}
\title{Compute S3}
\usage{
compute_s3(calculated_mass, experimental_mass, similarity_method)
}
\arguments{
\item{similarity_method}{Similarity method to be used}

\item{calculated_magnitude}{Calculated mass vector}

\item{experimental_magnitude}{Experimental mass vector}
}
\value{
Mass tolerance score S3
}
\description{
Returns the mass tolerance similarity score (S3)
}
