% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{exponential_decay_similarity}
\alias{exponential_decay_similarity}
\title{Exponential Decay Similarity}
\usage{
exponential_decay_similarity(a, b, method = "euclidian", normalize = T)
}
\arguments{
\item{a}{Vector a}

\item{b}{Vector b}

\item{method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}
}
\value{
Similarity between a and b
}
\description{
Determines the degree of similarity between a vector a and a vector b as 1/exp(d) where d corresponds to the distance between the two normalized vectors calculated according to the specified method
}
