% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_module.R
\name{export_mmass}
\alias{export_mmass}
\title{Export to mmass}
\usage{
export_mmass(pks_Matrix, matrix_annotation = rep(1,
  length(pks_Matrix$mass)), metadata = FALSE)
}
\arguments{
\item{pks_Matrix}{Peak matrix}

\item{matrix_annotation}{Vector determining whether each peak in the spectrum corresponds to the matrix or the tissue}

\item{metadata}{List containing metadata on the matrix annotation process}
}
\value{
None
}
\description{
Exports the results to a text file which can be read by mmass for easy interpretation and validation of the results.
}
