% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{find_first_lt}
\alias{find_first_lt}
\title{Find First LT}
\usage{
find_first_lt(x, k)
}
\arguments{
\item{x}{Input vector}

\item{k}{Threshold}
}
\value{
x with k precision
}
\description{
Returns the index of the first element in x lower than k
}
