% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{generate_before_after}
\alias{generate_before_after}
\title{Generate before and after files}
\usage{
generate_before_after()
}
\value{
None
}
\description{
Generate before and after zip files
}
