% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{generate_figures}
\alias{generate_figures}
\title{Generate PDF report}
\usage{
generate_figures(folder = "RESULTS")
}
\arguments{
\item{results}{Results matrix}
}
\value{
None
}
\description{
Generate PDF report
}
