% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{generate_file_name}
\alias{generate_file_name}
\title{Generate File Name}
\usage{
generate_file_name(base_name, extension = ".pdf", folder = "output/")
}
\arguments{
\item{base_name}{Base name including}

\item{extension}{File extension. Defaults to ".pdf"}

\item{folder}{Name of the folder in which to store the file. Defaults to "output/"}
}
\value{
File name that is not used in the
}
\description{
Generates a file name that has not yet been used
}
