% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{generate_pdf}
\alias{generate_pdf}
\title{Generate PDF report}
\usage{
generate_pdf(results, experiment_dir)
}
\arguments{
\item{results}{Results matrix}
}
\value{
None
}
\description{
Generate PDF report
}
