% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generate_pdf_report}
\alias{generate_pdf_report}
\title{Annotate matrix}
\usage{
generate_pdf_report(
  results,
  pks,
  full = NULL,
  filename = results$metadata$pks_name,
  folder = "output/",
  default_page_layout = NULL,
  include_summary = F
)
}
\arguments{
\item{pks}{Peak Matrix Image produced using rMSIproc}

\item{full}{Full spectrum before performing peak picking. It is used to give a higher degree of confidence to the S1 and S2 computation.}

\item{folder}{Name of the folder in which to store the pdf report}

\item{default_page_layout}{Page layout to be used in the pdf plotting.}

\item{include_summary}{Boolean value indicating if the summary is to be included in the pd}

\item{matrix_formula}{String giving the chemical formula of the matrix in the enviPat notation#'}

\item{s1_threshold}{Correlation between the theoretical and the real spectral pattern above which a given cluster is considered to be present}

\item{s2_threshold}{Correlation between the spatial images of the peak in a cluster above which the peaks are included in the gold truth (gt)}

\item{s3_threshold}{Correlation between the spatial images of the peak in a cluster above which the peaks are included in the gold truth (gt)}

\item{MALDI_resolution}{MALDI resolution that is used to merge nearby peaks in-silico as the equipment would in real life}

\item{tol_mode}{String determining the tolerance mode to be used when comparing the calculated masses with the experimental ones. "ppm": relative tolerance with respect to the calculated one in parts per million. "scans": Number of scans or datapoints present in the full spectrum -It is only applicable if full is provided.}

\item{tol_ppm}{Tolerance in parts per million. Only used if tol_mode="ppm".}

\item{tol_scans}{Tolerance in number of scans. Only used if tol_mode="scans".}

\item{mag_of_interest}{Magnitude of interest to use when performing the study. It can be "intensity" or "area".}

\item{normalization}{String indicating the normalization technique to use. Possible values: "None", "TIC","RMS", "MAX" or "AcqTic"}

\item{max_multi}{Maximum Cluster Multiplication. For each adduct -Adduct- it will generate a set of base formulas as follows: M+Adduct, 2M+Adduct, 3M+Adduct ... max_multi*M + Adduct}

\item{add_list}{List of adducts to be added to the matrix formula in the format defined by enviPat. Example: c("H1","Na1","K1")}

\item{sub_list}{List of compounds to be substracted to the matrix formula in the format defined by enviPat. Example: c("H1",H2O1")}

\item{generate_pdf}{Boolean indicating whether to generate a TIFF figures or not}

\item{similarity_method}{Similarity method to be used in the computation of the distance}

\item{pks_i}{Index of the peak matrix}

\item{plot_type}{String indicating the type of plot desired. Can have two possible values "debug" or "poster"}
}
\value{
Ground Truth: List of masses available in the image that correspond to the matrix.
}
\description{
Annotates the MS signals present in 'pks' related to the laser desorption/ionization promoting material used.
}
