% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning_module.R
\name{generate_synthetic_dataset}
\alias{generate_synthetic_dataset}
\title{Generate Synthetic Dataset}
\usage{
generate_synthetic_dataset(calculated_magnitude, image_correl)
}
\arguments{
\item{calculated_magnitude}{Calculated magnitude vector}

\item{image_correl}{Image correlation matrix of the experimental magnitude vector}
}
\value{
Spectral similarity score S2
}
\description{
Returns a synthetic dataset
}
