% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{get_closest_peak}
\alias{get_closest_peak}
\title{Get closest peak}
\usage{
get_closest_peak(calc_masses, experimental_masses)
}
\arguments{
\item{calc_masses}{Vector with calculated masses}

\item{experimental_masses}{Vector with experimental masses}
}
\value{
Peak matrix containing only the first image from pks
}
\description{
Returns the indices of "experimental_masses" that are the closest to each element in "calc_masses"
}
