% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{get_pareto_front}
\alias{get_pareto_front}
\title{Pareto Front}
\usage{
get_pareto_front(x, y, z)
}
\arguments{
\item{x}{Vector of first coordinates}

\item{y}{Vector of second coordinates}

\item{z}{Vector of third coordinates}
}
\value{
Boolean value indicating whether the experimental mass is within "tol_scans" of the calculated_mass
}
\description{
Returns the pareto front for a set of three dimensional points
}
