% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{ggplot_peak_image}
\alias{ggplot_peak_image}
\title{Plot peak image}
\usage{
ggplot_peak_image(pks, values = NA, title = "", isNA = F,
  chosen = T, in_pksMat = T)
}
\arguments{
\item{pks}{Peak matrix}

\item{values}{Values to be plotted}

\item{title}{Title of the plot}

\item{isNA}{Is NA}

\item{chosen}{Is chosen}

\item{in_pksMat}{Is in pksMat}
}
\description{
Plots a peak image
}
