% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{identify_matrix_correlation}
\alias{identify_matrix_correlation}
\title{Identify matrix correlation}
\usage{
identify_matrix_correlation()
}
\value{
None
}
\description{
Identifies de matrix peaks by following the method presented by Fonville et al. 2012.
}
\examples{
identify_matrix_correlation()

}
