% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pkg_opt}
\alias{pkg_opt}
\title{Verbose level global variable}
\usage{
pkg_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE,
  ADD = FALSE)
}
\arguments{
\item{...}{Options to change}

\item{RESET}{Reset the package options to the default}

\item{READ.ONLY}{Return only the READ.ONLY options}

\item{LOCAL}{Change to from global to local mode when TRUE and change from local to global mode when FALSE. In the local mode a copy of the options is created and the copy is modified.}

\item{ADD}{Add a new option on the fly}
}
\description{
1: Silent
 0: Application messages
-1: Code Sections
-2: Code subsections
-3: Debug mode
}
