% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(pks, normalize = TRUE, num_clusters = 2,
  correlation_order = 0)
}
\arguments{
\item{pks}{Peak Matrix}

\item{normalize}{Boolean value to determine if normalization is to be performed. If TRUE TIC normalization is performed}

\item{num_clusters}{Number of clusters}

\item{correlation_order}{Number of times that correlation needs to be applied.}
}
\value{
Nothing
}
\description{
Plot PCA
}
