% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_scores}
\alias{plot_scores}
\title{Compute scores}
\usage{
plot_scores(x, scores_list)
}
\arguments{
\item{x}{X variable}

\item{scores_list}{List of scores for each x}
}
\description{
Returns several scores for performance assessment of a given binary classification result.
}
