% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{plot_text}
\alias{plot_text}
\title{Plot text}
\usage{
plot_text(text, size = 3)
}
\arguments{
\item{text}{Text to be plotted}

\item{size}{Size of the text}
}
\description{
Plots text with white background
}
