% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_TIC}
\alias{plot_TIC}
\title{Plot image summary}
\usage{
plot_TIC()
}
\value{
None
}
\description{
Plot the eman spectra and a few representative images of the two MSI images used.
}
\examples{
plot_image_summary()

}
