% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{printv}
\alias{printv}
\title{Print verbose}
\usage{
printv(message, level)
}
\arguments{
\item{message}{String to print out}

\item{level}{Verbose level}
}
\value{
None
}
\description{
Print \code{message} if the verbose \code{level} is higher than the specified \code{threshold}.
}
\details{
1: Silent
 0: Application messages
-1: Code Sections
-2: Code subsections
-3: Debug mode
}
\examples{
printv("Hello World!",0)

}
