% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_module.R
\name{remove_matrix}
\alias{remove_matrix}
\title{Annotate matrix}
\usage{
remove_matrix(pks, results)
}
\arguments{
\item{pks}{Peak Matrix Image produced using rMSIproc}

\item{results}{Annotation results}
}
\value{
Ground Truth: List of masses available in the image that correspond to the matrix.
}
\description{
Annotates the MS signals present in 'pks' related to the laser desorption/ionization promoting material used.
}
