% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeMatrix_kMeansTranspose}
\alias{removeMatrix_kMeansTranspose}
\title{Remove Matrix k-Means Transpose}
\usage{
removeMatrix_kMeansTranspose(pks_Norharmane)
}
\arguments{
\item{pks_Norharmane}{Peak Matrix}
}
\value{
List of mass indices corresponding to the matrix peaks
}
\description{
Remove the matrix by performing k-means clustering on the transpose of the peak matrix.
The algorithm identifies similar spectral peaks.
}
