% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeMatrix_kMeansTransposeCor}
\alias{removeMatrix_kMeansTransposeCor}
\title{Remove Matrix k-Means Transpose based on correlation}
\usage{
removeMatrix_kMeansTransposeCor(pks_Norharmane)
}
\arguments{
\item{pks_Norharmane}{Peak Matrix}
}
\value{
None
}
\description{
Remove the matrix by performing k-means clustering on the transpose of the correlation of the peak matrix.
The algorithm identifies similar spectral peaks.
}
\examples{
removeMatrix_kMeansTransposeCor()

}
