% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeMatrix_padded}
\alias{removeMatrix_padded}
\title{Remove Matrix Padded}
\usage{
removeMatrix_padded(pks_Norharmane)
}
\arguments{
\item{pks_Norharmane}{Peak Matrix}
}
\value{
List of mass indices considered to be endogenous. The rest of the peaks are deamed as matrix related or non-anatomically relevant.
}
\description{
Remove the matrix following the method presented by Fonville et al. 2012.
It takes a padded image where there is a wide enough region of pixels outside of the tissue.
It consists of three basic steps.
}
