% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeMatrix_padded_new}
\alias{removeMatrix_padded_new}
\title{Remove Matrix Padded New}
\usage{
removeMatrix_padded_new(pks, normalize = TRUE, cor_threshold = 0.65,
  max_exo = 10)
}
\arguments{
\item{pks}{Peak Matrix}

\item{normalize}{Boolean value to determine if normalization is to be performed. If TRUE TIC normalization is performed}
}
\value{
List of mass indices considered to be endogenous. The rest of the peaks are deamed as matrix related or non-anatomically relevant.
}
\description{
Remove the matrix following the method presented by Fonville et al. 2012.
It takes a padded image where there is a wide enough region of pixels outside of the tissue.
It consists of three basic steps.
}
