% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specify decimal}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{Input double}

\item{k}{Precision}
}
\value{
x with k precision
}
\description{
Exports the results to a text file which can be read by mmass for easy interpretation and validation of the results.
}
