% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Generate report spectral clustering}
\usage{
spectral_clustering(pks, mass_range = c(min(pks$mass), max(pks$mass)),
  num_clus = 3, f = mean)
}
\arguments{
\item{pks}{Input Peak Matrix}

\item{mass_range}{Vector containing the minimum and maximum masses to be used in the process. The default is the complete mass range.}

\item{num_clus}{Number of clusters to use. Defaults to 3.}

\item{f}{Function to use when plotting the resulting spectral clusters in the spatial domain. The function should take a numerical vector and return a single number. Default is mean}
}
\value{
Matrix containing the masses processed along with the cluster to which each of them is assigned
}
\description{
Generates a pdf report with the results of a spectral clustering
}
