% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{tsne_before_after}
\alias{tsne_before_after}
\title{Generate before and after files}
\usage{
tsne_before_after(
  a = 1,
  max_iter = 5000,
  initial_dims = 100,
  pca = TRUE,
  perplexity = 20,
  eta = 100,
  theta = 0.5
)
}
\value{
None
}
\description{
Generate before and after zip files
}
