% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{v}
\alias{v}
\title{Verbose threshold:
0: Application messages
-1: Code Sections
-2: Code subsections
-3: Debug mode}
\format{An object of class \code{Verbose} (inherits from \code{Object}) of length 1.}
\usage{
v
}
\description{
Verbose threshold:
0: Application messages
-1: Code Sections
-2: Code subsections
-3: Debug mode
}
\keyword{datasets}
