% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute scores}
\usage{
compute_scores(gt, pos, neg, m)
}
\arguments{
\item{gt}{Ground truth: Positive values}

\item{pos}{Classified positives}

\item{neg}{Classified negatives}
}
\value{
List with all the computed scores. It includes F1 score and Brier score. As well as the number of
}
\description{
Returns several scores for performance assessment of a given binary classification result.
}
