% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_gt}
\alias{generate_gt}
\title{Generate gold standard}
\usage{
generate_gt(matrix_formula, pks, matching_method = "strict",
  cor_threshold = 0.6)
}
\arguments{
\item{matrix_formula}{String giving the chemical formula of the matrix in the enviPat notation}

\item{pks}{Peak Matrix Image}

\item{matching_method}{"loose": All peaks within a given tolerance are considered matrix peaks; "strict": Only peaks within a given tolerance that mantain the abundance ratios for each cluster are considered matrix peaks}

\item{cor_threshold}{Correlation between the theoretical and the real spectral pattern above which a given cluster is considered to be present and thus included in the gold truth (gt)}
}
\value{
Ground Truth: List of masses available in the image that correspond to the matrix.
}
\description{
Given the matrix formula and the matrix peaks present in the image it returns the ground truth defined as the list of peaks present in the image that correspond to the matrix.
}
