% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeMatrix_compareAu}
\alias{removeMatrix_compareAu}
\title{Remove Matrix Compare Au}
\usage{
removeMatrix_compareAu(pks_Norharmane, pks_Au)
}
\arguments{
\item{pks_Norharmane}{Peak Matrix}

\item{pks_Au}{Peak Matrix with gold coating}
}
\value{
None
}
\description{
Remove the matrix by comparing anatomically similar regions of the tissue to a reference image taken
with a gold matrix. Assesses the degree of similarity between peaks
It takes a padded image where there is a wide enough region of pixels outside of the tissue.
It consists of three basic steps.
}
