% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(
  matrix_formula = "Ag1",
  base_dirs = c("C:/Users/Gerard/Documents/1. Uni/1.5. PHD/images/Ag Software Test 1",
    "/home/gbaquer/msidata/Ag Software Test 1"),
  s1_threshold = 0.8,
  s2_threshold = 0.8,
  s3_threshold = 0.7,
  similarity_method = "euclidean",
  correlation_method = "pearson",
  experiment_name = "output",
  MALDI_resolution = 20000,
  tol_mode = "scans",
  tol_ppm = 1e-04,
  tol_scans = 4,
  mag_of_interest = "intensity",
  normalization = "None",
  max_multi = 10,
  add_list = NULL,
  sub_list = NULL,
  isobaric_detection = T,
  save_results = T,
  generate_pdf = T,
  generate_figures = F,
  default_page_layout = NULL,
  include_summary = F,
  dataset_indices = NULL
)
}
\arguments{
\item{dataset_indices}{Vector containing the indices of the datasets to be used}

\item{base_dir}{Base directory where the images are stored and the output reports should be stored}
}
\value{
None
}
\description{
Run a given experiment
}
