% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking_module.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{

  run_experiment(base_dir = "C:/Users/Gerard/Documents/1. Uni/1.5. PHD/images/Ag Software Test 1")
}
\value{
None
}
\description{
Run a given experiment
}
